Fix bug that launches same widget for different rdt label clicks
